feat: extract gateway_comments package (service + schemas)
Carves CommentsService (get_comments_for_post, get_comment_by_id) into a
new gateway_comments package, and moves CommentResponse/CommentsResponse
schemas out of gateway_oss, re-exporting them from gateway_oss for
backward compatibility. gateway_comments depends on gateway_core and
gateway_notes (the latter for the upcoming comments REST endpoint, which
returns a NoteResponse).
