+++ packages/gateway_comments/pyproject.toml
@@ -1,0 +1,16 @@
+[project]
+name = "gateway_comments"
+version = "3.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_notes", "pydantic>=2.13.4"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_comments"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_notes = { workspace = true }
